Add complete strategy patter example
In this example we have a duck simulator, with different types of ducks. Every duck can swim, quack, be displayed and fly.
Every duck swim the same, but each duck has a different display. Related with quack and fly, there are certain types of fly and quack options,
each duck belong to a specific group into these two fields.